Move the attachTagsToMemo method to MemoTagService.php
