+++ app/Services/TagServiceInterface.php
@@ -20,15 +20,4 @@ interface TagServiceInterface
      * @return bool
      */
     public function checkIfTagExists(int $userId, ?string $tagName): bool;
-
-    /**
-     * Attach tags to memo
-     *
-     * @param array|null $posts
-     * @param int $memoId
-     * @param bool $tagExists
-     * @param int $userId
-     * @return void
-     */
-    public function attachTagsToMemo(?array $posts, int $memoId, bool $tagExists, int $userId): void;
 }
