+++ tests/Unit/MemoTagServiceTest.php
@@ -5,6 +5,7 @@ namespace Tests\Unit;
 use Tests\TestCase;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use App\Repositories\MemoTagRepository;
+use App\Repositories\TagRepository;
 use App\Services\MemoTagService;
 use App\Models\MemoTag;
 use App\Models\Memo;
@@ -26,8 +27,9 @@ class MemoTagServiceTest extends TestCase
     public function setUp(): void
     {
         parent::setUp();
-        $memoTagService = new MemoTagRepository(new MemoTag());
-        $this->memoTagService = new MemoTagService($memoTagService);
+        $memoTagRepository = new MemoTagRepository(new MemoTag());
+        $tagRepository = new TagRepository(new Tag());
+        $this->memoTagService = new MemoTagService($memoTagRepository, $tagRepository);
     }
 
     /** @test */
@@ -48,6 +50,69 @@ class MemoTagServiceTest extends TestCase
         $this->assertSame($user->id, $memoTags[0]['user_id']);
         $this->assertSame($memo->id, $memoTags[0]['memo_id']);
         $this->assertSame($tag->id, $memoTags[0]['tag_id']);
+    }
+
+    /** @test */
+    public function it_can_new_tag_attach_tags_to_memo(): void
+    {
+        $user = User::factory()->create();
+        $memo = Memo::factory()->create(['user_id' => $user->id]);
+        $tag = Tag::factory()->create([
+            'name' => 'Test Old Tag',
+            'user_id' => $user->id
+        ]);
+
+        MemoTag::factory()->create([
+            'user_id' => $user->id,
+            'memo_id' => $memo->id,
+            'tag_id' => $tag->id,
+        ]);
+
+        $post = [
+            'content' => $memo->name,
+            'new_tag' => 'Test New Tag',
+        ];
+
+        $this->memoTagService->attachTagsToMemo($post, $memo->id, false, $user->id);
+
+        $this->assertDatabaseHas('memo_tags', [
+            'user_id' => $user->id,
+            'memo_id' => $memo->id,
+            'tag_id' => $tag->id,
+            'deleted_at' => null,
+        ]);
+    }
+
+    /** @test */
+    public function it_can_old_tag_attach_tags_to_memo(): void
+    {
+        $user = User::factory()->create();
+        $memo = Memo::factory()->create(['user_id' => $user->id]);
+        $tag = Tag::factory()->create([
+            'name' => 'Test Old Tag',
+            'user_id' => $user->id
+        ]);
+
+        MemoTag::factory()->create([
+            'memo_id' => $memo->id,
+            'tag_id' => $tag->id,
+            'user_id' => $user->id,
+        ]);
+
+        $post = [
+            'content' => $memo->name,
+            'new_tag' => null,
+            'tags' => ["$tag->id"]
+        ];
+
+        $this->memoTagService->attachTagsToMemo($post, $memo->id, true, $user->id);
+
+        $this->assertDatabaseHas('memo_tags', [
+            'user_id' => $user->id,
+            'memo_id' => $memo->id,
+            'tag_id' => $tag->id,
+            'deleted_at' => null,
+        ]);
     }
 
     /** @test */
